[Refactor] useCheckLogin custom hook 작성
